requirements.txt
selenium
beautifulsoup4
openai
python-dotenv
webdriver-manager